requirements.txt
face_recognition
cv2
numpy
xlwt
xlrd
xlutils.copy
PIL